Task "Refactor Daps V1 & V2" done
