+++ src/main/java/it/eng/idsa/businesslogic/service/impl/DapsServiceImpl.java
@@ -1,28 +1,17 @@
 package it.eng.idsa.businesslogic.service.impl;
 
 import java.io.IOException;
-import java.util.Date;
 
-import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.auth0.jwt.JWT;
-import com.auth0.jwt.algorithms.Algorithm;
-import com.auth0.jwt.exceptions.JWTDecodeException;
-import com.auth0.jwt.exceptions.SignatureVerificationException;
-import com.auth0.jwt.interfaces.DecodedJWT;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
-import it.eng.idsa.businesslogic.service.DapsService;
 import okhttp3.FormBody;
-import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.Response;
@@ -37,22 +26,12 @@ import okhttp3.Response;
 @ConditionalOnProperty(name = "application.dapsVersion", havingValue = "v1")
 @Service
 @Transactional
-public class DapsServiceImpl implements DapsService {
+public class DapsServiceImpl extends DapsServiceAbstractImpl {
 
 	private static final Logger logger = LoggerFactory.getLogger(DapsServiceImpl.class);
-
-	private String token = null;
 	
-	@Autowired
-	private DapsUtilityProvider dapsUtilityProvider;
-	
-	@Autowired
-	private OkHttpClient client;
-
-	@Value("${application.dapsUrl}")
-	private String dapsUrl;
-  
-	private String getJwTokenInternal() {
+	@Override
+	public String getJwTokenInternal() {
 
 		logger.debug("Get properties");
 
@@ -94,46 +73,6 @@ public class DapsServiceImpl implements DapsService {
 			if (responseDaps != null) {
 				responseDaps.close();
 			}
-		}
-		return token;
-	}
-
-	
-	@Override
-	public boolean validateToken(String tokenValue) {
-		boolean valid = false;
-		if(tokenValue==null) {
-			logger.error("Token is null");
-			return valid;
-		}
-		try {
-			DecodedJWT jwt = JWT.decode(tokenValue);
-			Algorithm algorithm = dapsUtilityProvider.provideAlgorithm(tokenValue);
-			algorithm.verify(jwt);
-			valid = true;
-			if (jwt.getExpiresAt().before(new Date())) {
-				valid = false;
-				logger.warn("Token expired");
-			}
-		} catch (SignatureVerificationException e) {
-			logger.info("Token did not verified, {}", e);
-		} catch (JWTDecodeException e) {
-			logger.error("Invalid token, {}", e);
-		}
-		return valid;
-	}
-
-
-	@Override
-	public String getJwtToken() {
-
-		token = getJwTokenInternal();
-
-		if (StringUtils.isNotBlank(token) && validateToken(token)) {
-			logger.info("Token is valid: " + token);
-		} else {
-			logger.info("Token is invalid");
-			return null;
 		}
 		return token;
 	}
