+++ src/main/java/it/eng/idsa/businesslogic/service/impl/DapsV2ServiceImpl.java
@@ -5,29 +5,18 @@ import java.security.KeyStoreException;
 import java.security.NoSuchAlgorithmException;
 import java.security.UnrecoverableKeyException;
 import java.security.cert.CertificateException;
-import java.util.Date;
 
-import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.auth0.jwt.JWT;
-import com.auth0.jwt.algorithms.Algorithm;
-import com.auth0.jwt.exceptions.JWTDecodeException;
-import com.auth0.jwt.exceptions.SignatureVerificationException;
-import com.auth0.jwt.interfaces.DecodedJWT;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.common.annotations.VisibleForTesting;
 
-import it.eng.idsa.businesslogic.service.DapsService;
 import okhttp3.FormBody;
-import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.Response;
@@ -39,22 +28,12 @@ import okhttp3.Response;
 @ConditionalOnProperty(name = "application.dapsVersion", havingValue = "v2")
 @Service
 @Transactional
-public class DapsV2ServiceImpl implements DapsService {
+public class DapsV2ServiceImpl extends DapsServiceAbstractImpl {
 	private static final Logger logger = LoggerFactory.getLogger(DapsV2ServiceImpl.class);
 
-	@Autowired
-	private OkHttpClient client;
-	
-	@Autowired
-	private DapsUtilityProvider dapsUtilityProvider;
-
-	private String token = null;
-
-	@Value("${application.dapsUrl}")
-	private String dapsUrl;
-
 	@VisibleForTesting
-	String getJwTokenInternal() {
+	@Override
+	public String getJwTokenInternal() {
 
 		Response jwtResponse = null;
 		try {
@@ -108,43 +87,5 @@ public class DapsV2ServiceImpl implements DapsService {
 			}
 		}
 		return token;
-	}
-
-	@Override
-	public boolean validateToken(String tokenValue) {
-		boolean valid = false;
-		if(tokenValue==null) {
-			logger.error("Token is null");
-			return valid;
-		}
-		try {
-			DecodedJWT jwt = JWT.decode(tokenValue);
-			Algorithm algorithm = dapsUtilityProvider.provideAlgorithm(tokenValue);
-			algorithm.verify(jwt);
-			valid = true;
-			if (jwt.getExpiresAt().before(new Date())) {
-				valid = false;
-				logger.warn("Token expired");
-			}
-		} catch (SignatureVerificationException e) {
-			logger.info("Token did not verified, {}", e);
-		} catch (JWTDecodeException e) {
-			logger.error("Invalid token, {}", e);
-		}
-		return valid;
-	}
-
-	@Override
-	public String getJwtToken() {
-
-		token = getJwTokenInternal();
-
-		if (StringUtils.isNotBlank(token) && validateToken(token)) {
-			logger.info("Token is valid: " + token);
-		} else {
-			logger.info("Token is invalid");
-			return null;
-		}
-		return token;
-	}
+	}	
 }
